Add unit tests and domain calculator class

+++ Solution/SynetecMvcAssessment/Models/BonusPoolCalculatorResultModel.cs
@@ -8,7 +8,7 @@ namespace InterviewTestTemplatev2.Models
     public class BonusPoolCalculatorResultModel
     {
         public Data.HrEmployee hrEmployee;
-        public int bonusPoolAllocation;
+        public decimal bonusPoolAllocation;
 
     }
 }
